backend/requirements.txt
fastapi
uvicorn
pydantic
google-cloud-bigquery
google-cloud-aiplatform
google-cloud-resource-manager
pandas
numpy
scikit-learn
google-generativeai
python-dotenv
python-multipart
db-dtypes
